Renamed PathInformation to PathDetail because the older name was just to long to type.
Changed the implementation in ModelBase.cs for the AsString and AsRawString properties, made then normal get, set wiht no backing property and no exception of the set as this is a abstract class so can very be set.

Changed Exception text to be more understandable, for exceptions where object instance is added yet not cloned from older used instance.