Found Bug and fixed. HL7Null was incorrectly reported as 'true' for a field when the first component of a field was HL7Null "" and yet component 2, 3, 4 ..etc were in fact populated. In this case The field's HL7Null should be false and component 1 of the field should be HL7Null = true. This same problem also cascaded down through Field , Component and Subcomponent in the same fashion.
Also, in fixing this bug it became clear that there is no such concept as HL7 Null for 'Content'. HL7Null or "" within content is actually just content. For example, consider this field: |""\H\This quoted text in bold\N\""|. It could be assumed that Content(0).IsHL7Null would equal 'true', yet this is not correct. This field, in plain text, would be ""This quoted text in bold"", yet in bold. Their simply in no concept of HL7Null in content unless that content is only HL7Null and nothing else e.g |""| or |^""| or |abc^def&""|. Oddly, this means in HL7 V2 there is no way to send "" as content only, it would have to have some other content or else it is interpreted as HL7 Null and in turn as as a delete content on file.